[run] don't remove files while application is equal

+++ packages/@weex/plugins/run/src/common/config.ts
@@ -57,8 +57,10 @@ class PlatformConfigResolver {
           fse.outputFileSync(file.replace(new RegExp(oldPath, 'ig'), newPath), data)
         })
       }
-      // remove old java source
-      fse.removeSync(path.join(basePath, javaSourcePath, 'java', oldPath))
+      if(oldPath !== newPath) {
+        // remove old java source
+        fse.removeSync(path.join(basePath, javaSourcePath, 'java', oldPath))
+      }
     },
   }
